creando interfaces para realizar las operaciones en sus respectivas clases
