+++ src/main/java/controllers/interfaces/passenger/ICreatePassengerDao.java
@@ -1,0 +1,7 @@
+package controllers.interfaces.passenger;
+
+import models.Passenger;
+
+public interface ICreatePassengerDao {
+    int createPassenger(Passenger passenger);
+}
