+++ src/main/java/controllers/interfaces/passenger/IEliminatedPassengerDao.java
@@ -1,0 +1,5 @@
+package controllers.interfaces.passenger;
+
+public interface IEliminatedPassengerDao {
+    
+}
