+++ src/main/java/controllers/interfaces/passenger/ISearchForIdentityDao.java
@@ -1,0 +1,13 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
+ */
+package controllers.interfaces.passenger;
+
+/**
+ *
+ * @author USUARIO
+ */
+public interface ISearchForIdentityDao {
+    
+}
