+++ src/main/java/controllers/interfaces/passenger/IShoPassengerSameDestinationDao.java
@@ -1,0 +1,8 @@
+package controllers.interfaces.passenger;
+
+import java.util.List;
+import models.Passenger;
+
+public interface IShoPassengerSameDestinationDao {
+    List<Passenger> showPassengersWithSameDestination(Passenger passenger);
+}
