+++ src/main/java/controllers/interfaces/passenger/IUpdatePassengerDao.java
@@ -1,0 +1,7 @@
+package controllers.interfaces.passenger;
+
+import models.Passenger;
+
+public interface IUpdatePassengerDao {
+    int updatePassenger(Passenger passenger);
+}
